requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.0
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12

pandas~=1.5.2
DeFiLlama~=1.1.0
APScheduler~=3.9.1.post1